Fix all AVR compilation warnings and errors
- Fix signed/unsigned comparison in LD2420.cpp:164
- Fix constructor member initialization order in LD245X.hpp
- Fix signed/unsigned comparison in Debug.hpp macros
- Fix format specifier for uint32_t (use %lu)
- Initialize val in CircularBuffer.hpp to avoid uninitialized warning
- Remove redundant HardwareSerial/SoftwareSerial includes from examples

Agent-Logs-Url: https://github.com/Fiooodooor/HLK-LD245X/sessions/fcd3163a-cb99-4369-95ad-07fcfe4cda6e

+++ src/LD245X.cpp
@@ -393,18 +393,18 @@ void LD245X::setNameString(const char* name)
     }
     else if(_sensorModel == SensorModel::LD2450) {
         snprintf(name_string, sizeof(name_string),
-                 "ld2450_%d", ObjectCounter::count());
+                 "ld2450_%lu", static_cast<unsigned long>(ObjectCounter::count()));
     }
     else if(_sensorModel == SensorModel::LD2451) {
         snprintf(name_string, sizeof(name_string),
-                 "ld2451_%d", ObjectCounter::count());
+                 "ld2451_%lu", static_cast<unsigned long>(ObjectCounter::count()));
     }
     else if(_sensorModel == SensorModel::LD2420) {
         snprintf(name_string, sizeof(name_string),
-                 "ld2420_%d", ObjectCounter::count());
+                 "ld2420_%lu", static_cast<unsigned long>(ObjectCounter::count()));
     } else {
         snprintf(name_string, sizeof(name_string),
-                 "ld245x_%d", ObjectCounter::count());
+                 "ld245x_%lu", static_cast<unsigned long>(ObjectCounter::count()));
     }
 }
